+++ nc_py_api/ex_app/providers/providers.py
@@ -2,6 +2,7 @@
 
 from ..._session import AsyncNcSessionApp, NcSessionApp
 from .speech_to_text import _AsyncSpeechToTextProviderAPI, _SpeechToTextProviderAPI
+from .text_processing import _AsyncTextProcessingProviderAPI, _TextProcessingProviderAPI
 
 
 class ProvidersApi:
@@ -9,9 +10,12 @@ class ProvidersApi:
 
     speech_to_text: _SpeechToTextProviderAPI
     """SpeechToText Provider API."""
+    text_processing: _TextProcessingProviderAPI
+    """TextProcessing Provider API."""
 
     def __init__(self, session: NcSessionApp):
         self.speech_to_text = _SpeechToTextProviderAPI(session)
+        self.text_processing = _TextProcessingProviderAPI(session)
 
 
 class AsyncProvidersApi:
@@ -19,6 +23,9 @@ class AsyncProvidersApi:
 
     speech_to_text: _AsyncSpeechToTextProviderAPI
     """SpeechToText Provider API."""
+    text_processing: _AsyncTextProcessingProviderAPI
+    """TextProcessing Provider API."""
 
     def __init__(self, session: AsyncNcSessionApp):
         self.speech_to_text = _AsyncSpeechToTextProviderAPI(session)
+        self.text_processing = _AsyncTextProcessingProviderAPI(session)